move Project and related struct out of mod.rs, clean up project mod api
