+++ src/huak/ops/install.rs
@@ -2,7 +2,7 @@ use crate::{
     env::venv::Venv,
     errors::{HuakError, HuakResult},
     package::python::PythonPackage,
-    project::{config::PythonConfig, Project},
+    project::{PythonConfig, Project},
 };
 
 /// Install all of the projects defined dependencies.
